+++ ui/server/graphql/resolvers/types/Transaction.ts
@@ -1,0 +1,6 @@
+export const __resolveType = (transaction) => {
+  if (transaction.bucketId) {
+    return "Contribution";
+  }
+  return "Allocation"; // GraphQLError is thrown
+};
