+++ ui/server/graphql/resolvers/types/index.ts
@@ -1,4 +1,16 @@
+import DateType from "./Date";
+
+export const Date = DateType;
 export * as InvitedMember from "./InvitedMember";
 export * as RoundMember from "./RoundMember";
 export * as GroupMember from "./GroupMember";
 export * as User from "./User";
+export * as Group from "./Group";
+export * as Round from "./Round";
+export * as Bucket from "./Bucket";
+export * as Transaction from "./Transaction";
+export * as Contribution from "./Contribution";
+export * as RoundTransaction from "./RoundTransaction";
+export * as Comment from "./Comment";
+export * as Flag from "./Flag";
+export * as CustomFieldValue from "./CustomFieldValue";
